Extend 'default' rake task with RuboCop linting
This adds a new 'lint' rake task that runs as part of a default call
to 'rake'. Having linting run locally by default makes it easier to
catch build issues, and is a prerequisite for moving to GitHub Actions,
since we want to avoid listing lots of separate build steps.

The 'lint' task runs RuboCop as a separate process, which is a technique
to avoid a LoadError when the gem is unavailable in production.