Fix concatenation Rubocop violation
The auto correct for this was a bit grim:

"[\"or\", [\"~\", [\"fact\", \"fqdn\"], \"^#{node_class}-\\\\d+\\\\.\"]]"

So I've used a percent literal for it.

It does come out with an unnerving and indecipherable amount of escape
slashes, but this does seem to be exactly what is needed for the tests
to continue passing.