Revise technical documentation section
This adds part of the standard intro we've used in other repos [1].
The `bin/...` command was wrong (there's no such directory), I was
also unable to create an -app stack in GOV.UK Docker, due to the
amount of hidden setup required for RabbitMQ (see below). In future
we may want to consider adding support for this.

Attempt at an -app stack:

  cache-clearing-service-app:
    <<: *cache-clearing-service
    depends_on:
      - rabbitmq
    environment:
      RABBITMQ_HOSTS: rabbitmq
      RABBITMQ_VHOST: /
      RABBITMQ_USER: guest
      RABBITMQ_PASSWORD: guest
      QUEUE: high
    command: bundle exec rake message_queue:consumer

Errors with the following:

  Bunny::NotFound: NOT_FOUND - no queue 'cache_clearing_service-high' in vhost '/'

We probably want something like the task we have in Publishing API [2].

[1]: https://github.com/alphagov/email-alert-api#technical-documentation
[2]: https://github.com/alphagov/publishing-api/blob/f3e6e7c3f665e675c72a174295c524e10ed01eff/lib/tasks/setup_exchange.rake