Clear the cache both with and without a GOV.UK account session
We've decided to implement account navigation on GOV.UK by
implementing a custom request header, which is set by Fastly if a
session cookie is present (and unset otherwise).  See [1].

So, since every page will potentially have two variants cached (when
we implement the new navigation), we will need the
cache-clearing-service to handle both of them.

For Fastly, set a session cookie; the `vcl_recv` method will then
produce the `GOVUK-Account-Session-Exists` header which the caching
uses.  For Varnish, set the `GOVUK-Account-Session-Exists` header
directly, as that assumes the request has come via Fastly.

I also changed the logger to print out the extra headers, if set, so
we'll see log entries like:

    Purging: https://www.gov.uk
    Purging: https://www.gov.uk {"Cookie"=>"__Host-govuk-account-session=1"}

[1] https://github.com/alphagov/govuk-cdn-config/pull/356