Fix logging.
Log to stdout when `RAILS_LOG_TO_STDOUT` is set (we're not using Rails
here, but introducing a new variable would be worse than this very
slight cognitive dissonance which has no practical downsides), otherwise
log to the same file we would have logged to before. (No bikeshedding
about boolean env var semantics, please. Checking for
not-the-empty-string is sufficient here because we're not designing a
public interface.)

We no longer do our own home-grown teeing of logs as this isn't needed
in production and all other use cases can be trivially handled by
piping through `tee` and/or tailing the log file.