Add retirement notice
## Context

Cache Clearing Service was build in August 2018. At the time GOV.UK pages would
typically return 30 minute cached version of a page, which was an unacceptably
long delay for publishers. The service listens for publishing events and purges
the registered routes from the Varnish and Fastly caches. This purging however
doesn’t remove all variants of the content and thus cannot be relied upon for
all circumstances. Consequently:
- any pages that are accessed with query strings or are within prefix routes
  won’t be purged from the cache by Cache Clearing Service.
- users could experience inconsistent experiences where one user can see a
  changed page and the other not
- GOV.UK HTTP responses incorrectly tell a client they can cache a page for the
  cache period when this isn’t strictly true

Content folks have been notified.

## Why

Currently, the cache period significantly reduced to 5 minutes.
[Purging a page from cache](https://docs.publishing.service.gov.uk/manual/purge-cache)
is documented in the dev docs.

For those reasons the service is deemed as redundant. Removing it will
reduce maintenance burden for developers and reduce cost of hosting the
service.